# Improve parallel scc, handle filtered multi-label graph