--- a/src/function/gds/component_ids.cpp
+++ b/src/function/gds/component_ids.cpp
@@ -6,20 +6,19 @@ namespace kuzu {
 namespace function {
 
 OffsetManager::OffsetManager(const table_id_map_t<offset_t>& maxOffsetMap) {
-     std::vector<table_id_t> tableIDVector;
-     for (auto [tableID, maxOffset] : maxOffsetMap) {
-         tableIDVector.push_back(tableID);
-     }
-     std::sort(tableIDVector.begin(), tableIDVector.end());
-     auto offset = 0u;
-     for (auto tableID : tableIDVector) {
-         tableIDToStartOffset.insert({tableID, offset});
-         offset += maxOffsetMap.at(tableID);
-     }
+    std::vector<table_id_t> tableIDVector;
+    for (auto [tableID, maxOffset] : maxOffsetMap) {
+        tableIDVector.push_back(tableID);
+    }
+    std::sort(tableIDVector.begin(), tableIDVector.end());
+    auto offset = 0u;
+    for (auto tableID : tableIDVector) {
+        tableIDToStartOffset.insert({tableID, offset});
+        offset += maxOffsetMap.at(tableID);
+    }
 }
 
-ComponentIDs ComponentIDs::getSequenceComponentIDs(
-    const table_id_map_t<offset_t>& maxOffsetMap,
+ComponentIDs ComponentIDs::getSequenceComponentIDs(const table_id_map_t<offset_t>& maxOffsetMap,
     const OffsetManager& offsetManager, storage::MemoryManager* mm) {
     auto result = ComponentIDs();
     for (auto [tableID, maxOffset] : maxOffsetMap) {
@@ -33,8 +32,8 @@ ComponentIDs ComponentIDs::getSequenceComponentIDs(
     return result;
 }
 
-ComponentIDs ComponentIDs::getUnvisitedComponentIDs(
-    const table_id_map_t<offset_t>& maxOffsetMap, storage::MemoryManager* mm) {
+ComponentIDs ComponentIDs::getUnvisitedComponentIDs(const table_id_map_t<offset_t>& maxOffsetMap,
+    storage::MemoryManager* mm) {
     auto result = ComponentIDs();
     for (auto [tableID, maxOffset] : maxOffsetMap) {
         result.denseObjects.allocate(tableID, maxOffset, mm);
@@ -45,7 +44,6 @@ ComponentIDs ComponentIDs::getUnvisitedComponentIDs(
     }
     return result;
 }
-
 
 bool ComponentIDsPair::update(offset_t boundOffset, offset_t nbrOffset) {
     auto boundValue = curData[boundOffset].load(std::memory_order_relaxed);
@@ -58,7 +56,8 @@ bool ComponentIDsPair::update(offset_t boundOffset, offset_t nbrOffset) {
     return false;
 }
 
-void ComponentIDsOutputVertexCompute::vertexCompute(offset_t startOffset, offset_t endOffset, table_id_t tableID) {
+void ComponentIDsOutputVertexCompute::vertexCompute(offset_t startOffset, offset_t endOffset,
+    table_id_t tableID) {
     for (auto i = startOffset; i < endOffset; ++i) {
         if (skip(i)) {
             continue;
@@ -70,5 +69,5 @@ void ComponentIDsOutputVertexCompute::vertexCompute(offset_t startOffset, offset
     }
 }
 
-}
-}
+} // namespace function
+} // namespace kuzu