… src/include/function/gds/gds_frontier.h
… src/include/function/gds/gds_state.h
… test/test_files/function/gds/scc.test