# Description

This PR contains the following changes

- make parallel SCC works for multi-labeled graph
- make parallel SCC works for filtered graph
- reuse WCC infrastructure because SCC is conceptually doing forward and backward WCC (coloring) twice
- improve parallel SCC performance because backward coloring only needs to be done on distinct colors.

|  | Master | This PR |
| -------- | ------- | ------- |
| Epinios| 100ms | 70ms |
| Gplus |  450ms | 120ms |


# Contributor agreement

- [x] I have read and agree to the [Contributor Agreement](https://github.com/kuzudb/kuzu/blob/master/CLA.md).